refactor(cli): collapse source_* re-export shells into _app (#1481 U7)
The four cli/services/source_{add,clean,content,wait}.py modules were pure
re-export shells over their transport-neutral _app/ cores (zero logic). Delete
them and repoint every consumer at _app directly:

- cli/source_cmd.py + cli/_source_render.py import from .._app.source_* (the
  source_add_service / source_clean_service module aliases are preserved, so the
  call-time looks_like_path / validate_upload_path / SourceAddValidationError /
  classify_junk_sources / build_source_add_plan attribute seams still resolve).
- Tests (test_source_add_service, test_cli_source_services,
  test_source_cmd_coverage) import from notebooklm._app.source_* instead.
- Drop the four GUARDED_PATHS inventory entries in test_services_boundary.py and
  the four cli/services/ tree rows in docs/architecture.md (the _app/ core rows
  for the same basenames stay; freshness check is bidirectional).

Behavior-preserving: --json bytes unchanged (ADR-0015); no cassette re-records
(ADR-0021). Unit + cli_vcr suites green; ruff/mypy clean.
